requirements_simple.txt
python-fasthtml
uvicorn
plotly
folium
pandas
numpy
lightgbm
scikit-learn
holidays
geopandas
